+++ src/pages/RepairJob/RepairJobDeliveryTicket.tsx
@@ -11,7 +11,7 @@ import { AiFillFilePdf } from "react-icons/ai";
 import axiosInstance from "../../axios/axiosInstance";
 import { CustomToastContext } from "../../StateProvider/CustomToastContext/CustomToastContext";
 import NoDataCell from "../../components/Helpers/NoDataCell";
-import { deliveryTicket, gridLoadingTimeout, rentalManagement, repairJob, repairJobStatus, sidebarResource } from "../../constants/helpers";
+import { deliveryTicket, gridLoadingTimeout, rentalManagement, repairJob, repairJobStatus, sidebarResource, INVENTORY_STATUS } from "../../constants/helpers";
 import ConfirmationDialog from "../../components/Helpers/ConfirmationDialog";
 import { groupBy } from 'lodash';
 import AddBoxRoundedIcon from '@material-ui/icons/AddBoxRounded';
@@ -112,7 +112,7 @@ const RepairJobDeliveryTicket = ({ repairJobData, setNextButtonDisabled, setPrev
 
             tempProductInventory.forEach((d) => {
               d["_id"] = d["id"];
-              d["hideSelection"] = d.status === "In-Transit"; // || (d.hasOwnProperty("isDelivered") && d["isDelivered"] === true) || d["repaired"];
+              d["hideSelection"] = d.status === INVENTORY_STATUS.indTransit; // || (d.hasOwnProperty("isDelivered") && d["isDelivered"] === true) || d["repaired"];
             })
 
             if (repairJobData["typeOfRepair"] === "Internal" && repairJobData["plant"]?.optionValue !== repairJobData["repairPlant"]?.optionValue && tempProductInventory.some(s => s["repaired"] === true)) {
@@ -295,7 +295,7 @@ const RepairJobDeliveryTicket = ({ repairJobData, setNextButtonDisabled, setPrev
       {
         repairJobData && repairJobData["status"] !== repairJobStatus[2] &&
         <IconButton
-          disabled={selectedRecords.length === 0 || selectedRecords.some(f => f.hasOwnProperty("isDelivered")) || selectedRecords.some(f => !f.hasOwnProperty("deliveryTicketId")) || selectedRecords.some(f => f.repaired === true)}
+          disabled={selectedRecords.length === 0 || selectedRecords.some(f => f.hasOwnProperty("isDelivered") && f.isDelivered === true) || selectedRecords.some(f => f.repaired === true)}
           onClick={() => {
             setShowRemoveAssetFromLoadingTicketDialog(true)
           }}